+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 4.6.4 ##
+
+1.  UI Toolkit: Fix `double` & `float` edting round [#215](https://github.com/TylerTemp/SaintsField/issues/215)
+2.  Fix error should be dismissed when user override the `Equal` method which raises an error
+3.  UI Toolkit: Fix `ShowInInspector` an error when add element to a list/array
+4.  UI Toolkit: `ShowInInspector` now only show `public` `instance` field to avoid loop calling
+
 ## 4.6.3 ##
 
 Fix broken auto getters since 4.6.2... (sad face)